+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # CHANGELOG
 
 ## [Unreleased]
+### Added
+* Added `Kreait\Firebase\Auth::getUsers()` enables retrieving multiple users at once.
+  ([#477](https://github.com/kreait/firebase-php/pull/477))
+  ([Documentation](https://firebase-php.readthedocs.io/en/latest/user-management.html#get-information-about-multiple-users))
 
 ## [5.9.0] - 2020-10-04
 ### Added